Purpose: porting to T3E Solution: netcdf Fortran API tests failed because test program is using REAL and INTEGER for nclong and ncfloat datatype. T3E Fortran REAL and INTEGER are both 8 bytes while C float and int are 4 bytes.  Changed them to REAL*4 and INTEGER*4 to be compatible.  Though REAL*4 and INTEGER*4 are not pure F77 but most compilers support that extension.

+++ mfhdf/fortran/hdftest.f
@@ -1196,14 +1196,14 @@ C
 C
 C  Data declaration
 C
-         integer   data(n,m),
+         integer*4 data(n,m),
      .             chunk(nc,mc),
      .             chunk_out(nc,mc),
      .             chunk_data(nc,mc,n_nc,n_mc)
 C
 C  Buffers to hold part of the data when we read data back
 C
-         integer    part_data(n_part,m_part)
+         integer*4 part_data(n_part,m_part)
 C
 C  HDF parameters initialization
 C
@@ -2197,11 +2197,11 @@ C
 C
 C  Data declaration
 C
-         integer   data(n,m)
+         integer*4 data(n,m)
 C
 C  Buffer to hold part of the data when we read data back
 C
-         integer    part_data(n_part,m_part)
+         integer*4 part_data(n_part,m_part)
 C
 C  HDF parameters initialization
 C
